--- tower-http/src/services/fs/tests/mod.rs
@@ -1,2 +1,0 @@
-mod serve_dir;
-mod serve_file;
